Switched user to ObjectID identification instead of string

+++ src/repos/comments-repo.ts
@@ -1,10 +1,7 @@
 import {CommentModelClass} from "../db/db";
-import {
-    CommentDbModel,
-    CommentViewModel
-} from "../models/Comments/CommentModel";
+import {CommentDbModel, CommentViewModel} from "../models/Comments/CommentModel";
 import {getCommentViewModel} from "../helpers/map-CommentViewModel";
-import {ObjectId} from "mongodb";
+import {createObjectIdFromSting} from "../helpers/map-ObjectId";
 
 
 export const commentsRepo = {
@@ -16,22 +13,16 @@ export const commentsRepo = {
         //     }
         // })
         // return result.matchedCount === 1
-        try {
-            // Отлавливаем ошибку, когда в ID передается неверные данные
-            // Argument passed in must be a string of 12 bytes or a string of 24 hex characters or an integer
-            // BSONError: Argument passed in must be a string of 12 bytes or a string of 24 hex characters or an integer
-            const _id = new ObjectId(id);
-            const commentInstance = await CommentModelClass.findOne({"_id": _id})
-            if (!commentInstance) return false
+        const _id = createObjectIdFromSting(id)
+        if (_id === null) return false
+        const commentInstance = await CommentModelClass.findOne({"_id": _id})
+        if (!commentInstance) return false
 
-            commentInstance.content = content
+        commentInstance.content = content
 
-            await commentInstance.save()
+        await commentInstance.save()
 
-            return true
-        } catch (e) {
-            return false
-        }
+        return true
 
     },
     async deleteComment(id: string): Promise<boolean> {
@@ -40,21 +31,18 @@ export const commentsRepo = {
         // let result = await CommentModelClass.deleteOne({"id": id})
         // return result.deletedCount === 1
 
-        try {
-            const _id = new ObjectId(id);
-            const commentInstance = await CommentModelClass.findOne({"_id": _id})
-            if (!commentInstance) return false
+        const _id = createObjectIdFromSting(id)
+        if (_id === null) return false
+        const commentInstance = await CommentModelClass.findOne({"_id": _id})
+        if (!commentInstance) return false
 
-            await commentInstance.deleteOne()
+        await commentInstance.deleteOne()
 
-            return true
-        } catch {
-            return false
-        }
+        return true
 
 
     },
-    async createComment(newComment: CommentDbModel ): Promise<CommentViewModel> {
+    async createComment(newComment: CommentDbModel): Promise<CommentViewModel> {
         // Mongo native driver code
         // await CommentModelClass.insertMany([newComment])
